Cleaned up method call code.

+++ TightWiki.Shared/Models/WeightedToken.cs
@@ -5,7 +5,7 @@ namespace TightWiki.Shared.Models
     public class WeightedToken
     {
         public string Token { get; set; }
-        public int Weight { get; set; }
+        public double Weight { get; set; }
         public string DoubleMetaphone { get; set; }
 
         public PageToken ToPageToken(int pageId)
